+++ handler/put.go
@@ -24,123 +24,39 @@
 package handler
 
 import (
-	"bytes"
-	"io"
 	"net/http"
 	"strings"
 
-	sf "github.com/gogufo/gufo-api-gateway/gufodao"
 	pb "github.com/gogufo/gufo-api-gateway/proto/go"
 
 	"github.com/spf13/viper"
 )
 
 func ProcessPUT(w http.ResponseWriter, r *http.Request, t *pb.Request, version int) {
-
 	path := r.URL.Path
 	patharray := strings.Split(path, "/")
-	pathlenth := len(patharray)
-	//p := bluemonday.UGCPolicy()
-
-	if pathlenth < 3 {
-		errorAnswer(w, r, t, 401, "0000235", "Wrong Path Lenth")
+	if len(patharray) < 3 || *t.Module == "entrypoint" {
+		errorAnswer(w, r, t, 401, "0000235", "Wrong Path Length")
 		return
 	}
 
-	if *t.Module == "entrypoint" {
-		errorAnswer(w, r, t, 401, "0000235", "Wrong Path Lenth")
-		return
-	}
-
-	//check for session
+	// 🔐 Check session
 	if viper.GetBool("server.session") {
 		t = checksession(t, r)
-
 		if t.UID != nil && *t.Readonly == int32(1) {
-
 			errorAnswer(w, r, t, 401, "0000235", "Read Only User")
 			return
-
 		}
 	}
 
 	vrs := "v3"
-	t.APIVersion = &vrs
-
 	if version == 2 {
 		vrs = "v2"
-		t.APIVersion = &vrs
 	}
+	t.APIVersion = &vrs
 
-	args := make(map[string]interface{})
-	for key, value := range r.Form {
+	param := "stream"
+	t.IR = &pb.InternalRequest{Param: &param}
 
-		if len(value) == 1 && len(value) != 0 {
-			args[key] = value
-		}
-	}
-
-	t.Args = sf.ToMapStringAny(args)
-
-	/*
-		var (
-			buf        []byte
-			firstChunk bool
-		)
-	*/
-	//PUT mean file upload, so we check for file data
-	file, handler, err := r.FormFile("file")
-
-	if err != nil || file == nil || handler.Filename == "" {
-		errorAnswer(w, r, t, 400, "0000235", "Missing File")
-		return
-	}
-
-	defer file.Close()
-
-	buft := bytes.NewBuffer(nil)
-	if _, err := io.Copy(buft, file); err != nil {
-
-		errorAnswer(w, r, t, 400, "0000235", err.Error())
-		return
-	}
-
-	t.Filename = &handler.Filename
-	t.File = buft.Bytes()
-
-	//start uploader
-	/*
-		buf = make([]byte, chunkSize)
-		firstChunk = true
-		for {
-			n, errRead := file.Read(buf)
-			if errRead != nil {
-				if errRead == io.EOF {
-					errRead = nil
-					break
-				}
-				errorAnswer(w, r, t, 400, "0000235", "errored while copying from file to buf")
-				return
-			}
-
-			if firstChunk {
-				t.Filename = &handler.Filename
-				t.File = buf[:n]
-				firstChunk = false
-			} else {
-				t.File = buf[:n]
-			}
-			if err != nil {
-				errorAnswer(w, r, t, 400, "0000235", "failed to send chunk via stream file")
-				return
-			}
-
-		}
-	*/
-
-	//Check is it plugin or GRPC server
-
-	//Load microservice
 	connectgrpc(w, r, t)
-
 }