+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Check Session improvememts
 - TLS/mTLS + connection pool
 - add timeout for grpc connections
+- add multifile upload
 
 ##.1.17.1
 - Update Golang Version to 1.23